+++ tybotRoute/models/IntentFormStudy.js
@@ -11,6 +11,16 @@ class IntentForm {
     this.CURRENT_FIELD_INDEX_KEY = "tilebot:requests:" + this.requestId + ":currentFieldIndex"
     this.CURRENT_FORM_KEY = "tilebot:requests:" + this.requestId + ":currentForm"
     this.log = options.log;
+    this.requestParameters = options.requestParameters;
+  }
+
+  getParam(paramKey) {
+    if (!this.requestParameters) {
+      return null;
+    }
+    console.log("this.requestParameters:", this.requestParameters);
+    console.log("this.requestParameters[" + paramKey + "]:", this.requestParameters[paramKey]);
+    return this.requestParameters[paramKey];
   }
 
   async setValue(key, value) {
@@ -79,18 +89,26 @@ class IntentForm {
       if (this.log) {console.log("current_field is undefined")}
       current_field = 0;
       
+      // first "freeze" the currente form, so that eventual form modifications
+      // do not compromise this form processing.
       await this.setValue(this.CURRENT_FORM_KEY, JSON.stringify(this.form));
+
     //   if (getParam(this.form.fields[current_field].name)) {
     //     //current_field++;
     //     await this.setValue(this.CURRENT_FIELD_INDEX_KEY, current_field); //=0
     //     return this.getMessage(current_value)
     //   }
-      
+      // set the first field under the "await the response-value" state (=0)
       await this.setValue(this.CURRENT_FIELD_INDEX_KEY, current_field);
-      const is_current_value = getParam(this.form.fields[current_field].name);
+      // now look for an already set value for this field in request parameters
+      console.log("params", this.requestParameters);
+      console.log("checking field:", this.form.fields[current_field].name);
+      const is_current_value = this.getParam(this.form.fields[current_field].name);
       if (is_current_value) {
+        if (this.log) {console.log("is_current_value!", is_current_value);}
         return await this.getMessage(is_current_value);
       }
+      if (this.log) {console.log("Form asking fist value. No 'is_current_value' for first form field", is_current_value);}
       if (this.log) {console.log("INTENT_FORM:", this.form);}
       if (this.log) {console.log("CURRENT FIELD:", current_field);}
       let message = {
@@ -166,16 +184,20 @@ class IntentForm {
       else {
         if (this.log) {console.log("Processing next field:", current_field)}
         await this.setValue(this.CURRENT_FIELD_INDEX_KEY, current_field);
-        const is_current_value = getParam(this.form.fields[current_field].name);
+
+        console.log("params", this.requestParameters);
+        console.log("checking field:", this.form.fields[current_field].name);
+        
+        const is_current_value = this.getParam(this.form.fields[current_field].name);
         if (is_current_value) {
+          if (this.log) {console.log("is_current_value!", is_current_value);}
             return await this.getMessage(is_current_value);
         }
         else {
-            let message = {
-              text: current_form.fields[current_field].label
-            }
             return {
-              message: message
+              message: {
+                text: current_form.fields[current_field].label
+              }
             };
         }